Move App back to src root
